검토 완료"

+++ src/main/java/com/oracle/coffee/dto/km/MfgRpDTO.java
@@ -1,5 +1,6 @@
 package com.oracle.coffee.dto.km;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import org.springframework.format.annotation.DateTimeFormat;
@@ -15,8 +16,8 @@ public class MfgRpDTO {		// 생산보고
 	private LocalDateTime	mfg_end_date;		// 실제 생산 완료 날짜  (수불마감 시 필요)
 	private int 			mfg_mat;			// 실제 투입 원재료 총중량(g)
 	private int 			mfg_end;			// 실제 생산 완료  총중량(g)
-	private double 			yield;				// 실제 수율(%)
-	private double 			pct;				// 펀차(소수점 첫째 자리)
+	private BigDecimal 		yield;				// 실제 수율(%)
+	private BigDecimal 		pct;				// 펀차(소수점 첫째 자리)
 	private String 			note;				// 비고
 	private int 			mfg_rp_reg_code;	// 사원번호(신청인과 다를 수 있다)
 	private LocalDateTime 	report_reg_date;	// 등록일
